better folder structure and cleanup of unused types and queries
