+++ src/groups/RouteNetworkDiagram/EditDiagram/EditDiagramGql.ts
@@ -1,12 +1,3 @@
-export interface DiagramQueryResponse {
-  schematic: {
-    buildDiagram: {
-      envelope: Envelope;
-      diagramObjects: Diagram[];
-    };
-  };
-}
-
 export interface Envelope {
   minX: number;
   maxX: number;
@@ -26,65 +17,6 @@ export interface Geometry {
   type: string;
   coordinates: string;
 }
-
-export const GET_DIAGRAM = `
-query ($routeNetworkElementId: ID!) {
-  schematic {
-    buildDiagram(
-      routeNetworkElementId: $routeNetworkElementId
-    ) {
-      envelope {
-        minX
-        maxX
-        minY
-        maxY
-      }
-      diagramObjects {
-        refId
-        refClass
-        geometry {
-          type
-          coordinates
-        }
-        style
-        label
-      }
-    }
-  }
-}
-`;
-
-export interface DiagramUpdatedResponse {
-  schematicDiagramUpdated: {
-    diagramObjects: Diagram[];
-    envelope: Envelope;
-  };
-}
-
-export const SCHEMATIC_DIAGRAM_UPDATED = `
-subscription ($routeNetworkElementId: ID!) {
-  schematicDiagramUpdated(
-    routeNetworkElementId: $routeNetworkElementId
-  ) {
-    envelope {
-      minX
-      maxX
-      minY
-      maxY
-    }
-    diagramObjects {
-      refId
-      refClass
-      geometry {
-        type
-        coordinates
-      }
-      style
-      label
-    }
-  }
-}
-`;
 
 export interface AffixSpanEquipmentParams {
   spanSegmentId: string;
