+++ src/groups/RouteNetworkDiagram/EditDiagram/index.ts
@@ -1,0 +1,2 @@
+import EditDiagram from "./EditDiagram";
+export default EditDiagram;
