+++ src/groups/RouteNetworkDiagram/FeatureDetail/FeatureDetail.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+FeatureDetail
+
+function FeatureDetail() {
+
+}
+
+export default FeatureDetail;
