+++ src/groups/RouteNetworkDiagram/ReadOnlyDiagram/index.ts
@@ -1,0 +1,2 @@
+import ReadOnlyDiagram from "./ReadOnlyDiagram";
+export default ReadOnlyDiagram;
