--- src/groups/RouteNetworkDiagram/RouteNetworkDiagram.scss
@@ -1,3 +1,0 @@
-.route-network-diagram {
-  width: 100%;
-}
