+++ src/groups/RouteNetworkDiagram/index.ts
@@ -1,4 +1,3 @@
 import DiagramWrapper from "./DiagramWrapper";
-import "./RouteNetworkDiagram.scss";
 
 export default DiagramWrapper;
